+++ lib/daisybill_api/models/attachment.rb
@@ -10,6 +10,7 @@ module DaisybillApi
         inactive: :boolean,
         review_status: :string,
         review_errors: [:string],
+        review_warnings: [:string],
         document_file_name: :string,
         document_content_type: :string,
         document_file_size: :integer,